requirements.txt
openai
langchain
python-dotenv
streamlit
langchain_community